+++ src/RareItemSet.cc
@@ -196,7 +196,7 @@ RareItemSet::ParsedRELData::ParsedRELData(const SpecCollection& collection) {
   if (collection.box_specs.size() > 0xFF) {
     throw std::runtime_error("area_norm value too high");
   }
-  for (uint8_t area_norm = 0; area_norm < collection.box_specs.size(); area_norm++) {
+  for (size_t area_norm = 0; area_norm < collection.box_specs.size(); area_norm++) {
     for (const auto& spec : collection.box_specs[area_norm]) {
       uint8_t area_norm_plus_1 = area_norm + 1;
       this->box_rares.emplace_back(BoxRare{.area_norm_plus_1 = area_norm_plus_1, .drop = spec});
